Make options symbols.
This might be a bad idea. But for now it will eventually allow me to
look up in string named tuples because there is no way to convert a
string to a symbol, however you can turn a symbol into a string. So when
in doubt, seems like maybe I should default to symbols. Or this is a
really bad idea. Some time in the future I might find this commit
message and hate myself. Only time will tell.